Filter and merge tractors
We should only keep reference to the last position
of every driver to avoid having to iterate through
all of them every time we need to update tractors.

A map of every driver should be store also, so we
don't lose reference to them even if in the next
request a driver doesn't come up in the json we can
still have a reference from it.